requirements.txt
faiss-cpu
nltk
sentence-transformers
joblib
llama-cpp-python
fastapi
pydantic
uvicorn